requirements.txt
pillow
opencv-python
matplotlib
imageio
scipy
scikit-image
imread
pytest
